Create book form to allow user input
Created "new-book" button that will display the form on-click, a "close" button that will hide the form on-click, and a form fieldset that allows the user to specify the title, author, number of pages, and
whether or not the book is read. Finally, an "add-book" button was created that will add the book to the library.