Create removeBook method on Book and run it on button 'click' event
The removeBook method removes the book element from the DOM and the Book object from the array simultaneously, then loops over each Book object of the libraryArray array in order to update the dataset.count value of each book element that corresponds to the iterated Book object. This ensures that the index of book elements remains equal with the index of Book objects.