Create renderDeleteButton(), remove(), and removeBook() methods
The renderDeleteButton and remove methods are created on the Book class.
renderDeleteButton clones a deleteButton that already exists on the DOM,
but is hidden, and binds to it a 'click' event that calls the remove()
method.
When remove() method is called, it further calls the removeBook() method
created on the Library class, passing the current this.#container as
argument. This gives the Library access to the otherwise private
book DOM element while also having access to the Book object stored
in the books array.
The removeBook() method associates the book element with the Book object
using the 'count' dataset of the book element as index, then
removes both the book DOM element and the Book object from their
respective containers.