Create getter for #container and call all render methods in render()
A getter for #container is necessary to allow the removeBook() method
on the Library class to access it.
The render() method of Book's class now acts as a main function that
renders the book container and then further calls all render methods
(renderTitle, renderDeleteButton...). This allows for all instances
of the Book class to render their values with one method.